Profile Implementation and general clean up

+++ cryptovip/Models/PaymentOptionModel.cs
@@ -2,7 +2,7 @@
 {
     public class PaymentOptionModel
     {
-        public int Id { get; internal set; }
+        public byte Id { get; internal set; }
 
         public string Currency { get; internal set; }
